Problem 3, attempt.
As I've been coding, there donned on me that there is probably a better way to solve the problem by using further decomposition (i.e., in this case, add methods to lay down bricks west, and lay down bricks east - as sub-methods). There is another (potentially slicker) way to solve the problem using no switch statements (like I should have to begin with), by using the bot's location % 2 to see if it's on a proper square.