Fixed problem 3.
Final modification to problem 3. The 7x7 board was pretty challenging, and I've definitely diverged from the recommendation that the class has made that each method be at most about 15 lines. My layDownBeepers() method is almost 90 lines long. There is definitely room to improve on this exercise, especially in light of the fact that I'm using java constructs (such as a switch statement) to solve the problem, when that methodology hasn't been introduced into the class yet...maybe at a later date I'll revisit this one.