Problem 4; solved.
Refactored the last bit of code to get Karel to seek the edge beepers. This logic is slightly different, in that I separated the 'find edge beeper' and 'test for last beeper' into separate logic - which makes much more sense. Now Karel is able to get himself to the edges of the remaining beepers, check to see if it is the last beeper in the world, and take the appropriate action from that point.