.class files change.
No idea why these files are showing up as changed in the index since the .gitignore file is supposed to ignore all *.class files. The only thing I did differently was change the build environment for a different project 'Assignment2'...not sure why that would have had this effect, though.