Fixed Pyramid's missing .class.
I encountered an error on a subsequent commit wherein the .class file was deleted and then the program (and all my code) would no longer compile and work in eclipse. Now, I've gone back and been able to re-capture the .class file (from when it was working) as well as the changes in my program. Next step is to merge this back into the master branch.