Incomplete Pyramid.
The Pyramid problem is incomplete, but it's because I think there is an issue with some of the library calls or Eclipse set upt that I'm working with. Nothing I do seems to change the placement of the bricks...of course, I haven't experimented with just placing the bricks according to a numeric scheme (i.e. hard-coding them to see what happens...maybe later).