Breakout - add Ball and GameOver functions.
Functions to add the player ball, as well as display a game over message after the player has had three turns. There is some issue here with the mouseClicked() event function that I don't quite understand yet, so I'll have to hit the book, but so far, it's running pretty smoothly all the way through, though the ball needs motion and bouncing next.