Breakout - Collision logic re-factoring.
Re-factored the code that checks whether or not the ball
has hit an edge of the world. This will open up a nice method
for testing whether or not a collision happends with other
game entities, as well - such as the paddle, and eventually
the bricks. I also chose to add an initial Y_VEL static constant
in order to ensure that between turns, the ball will start
downward. Previously, when the ball would hit the bottom, the
velocity would switch directions - of course, that code could
just as easily be changed in the future.