Hangman: Part III
Implements part 3 of the problem. Adds a new file reader (BufferedReader),
that will read in a list of words and also modifies the structure the words
are contained in - uses an ArrayList. Further modifies the two functions
that operate on this new structure. Replaces the earlier stubs that were used
in the original program to get it functioning properly.